// pages/types/rescueTeams.type.ts
import { Model } from "./model.type";

export type RescueTeamType = Model & {  
    id: "6a0aeb02-f0ec-4b22-986e-7d9505a265fe",
    name: string;
    email: string;
    phoneNumber: string;
    rescueTeamCategory: string;
    createdAt: "2023-10-14T10:15:08.126Z",
    updatedAt: "2023-10-14T10:15:08.126Z",
    status: string;
    visibility: string;

};